requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blessed==1.20.0
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==4.4.2
distro==1.9.0
editor==1.6.6
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inquirer==3.4.0
jiter==0.8.2
moviepy==2.1.2
numpy==2.2.2
openai==1.62.0
opencv-python==4.11.0.86
pillow==10.4.0
proglog==0.1.10
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
python-dotenv==1.0.1
python-ffmpeg==2.0.12
readchar==4.2.1
requests==2.32.3
runs==1.2.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
xmod==1.8.1
